feat(events): change z index of sentry feedback button instead of hiding it
